Configure the DB sandbox properly for integration testing
Extend the item controller to use the DB. This makes it possible to
write integration tests which depend on the DB. The sandbox adapter is
configured properly now.

Unfortunately I had to disable the async feature for the integartion
tests, because we want the API calls to see the changes in the DB done
in the test.